% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{calc}
\alias{calc}
\title{Programmatically Calculate New Columns}
\usage{
calc(df, labels, formulas, prefix = "")
}
\arguments{
\item{df}{The data frame you're manipulating. The \code{formulas} should be based off of the column names of df.}

\item{labels}{A character vector used as the names for the newly created columns.}

\item{formulas}{A character vector of formulas (e.g. \code{col1 / col2}) used to generate the new columns.}

\item{prefix}{A prefix to append to the beginning of all of the column names in \code{labels}.}
}
\value{
A data frame with the newly calculated columns.
}
\description{
Create new columns using a list of column names (labels) and formulas. \code{labels} and \code{formulas} should both be character vectors of the same length
}
\examples{
calc(mpg, "average mpg", "mean(c(cty, hwy), na.rm = TRUE)")
}
